requirements.txt
google-cloud
google-cloud-vision
google-cloud-bigquery
google-cloud-translate
wget